Finish implementation of channel flags
