+++ functions/src/aqi-calculation/invalid-aqi-errors.ts
@@ -1,33 +1,4 @@
 /**
- * Interface for the structure of a sensor's data, used in the `current-reading`
- * collection in the `sensors` doc.
- * - `purpleAirId` - PurpleAir sensor ID
- * - `name` - PurpleAir sensor name
- * - `latitude` - latitude of sensor
- * - `longitude` - longitude of sensor
- * - `isValid` - if the current NowCast PM2.5 and AQI value are valid
- * - `isActive` - if we should be actively gathering data for the sensor
- * - `aqi` - the current AQI for the sensor, or `NaN` if not enough valid data
- * - `nowCastPm25` - the current NowCast corrected PM2.5, or `NaN` if not enough valid data
- * - `readingDocId` - document ID of the for the sensor in the sensors collection in Firestore
- * - `lastValidAqiTime` - the last time the AQI was valid, or null if unknown
- * - `lastSensorReadingTime` - the last time the sensor gave a reading, or null if unknown
- */
-interface CurrentReadingSensorData {
-  purpleAirId: number;
-  name: string;
-  latitude: number;
-  longitude: number;
-  isValid: boolean;
-  isActive: boolean;
-  aqi: number;
-  nowCastPm25: number;
-  readingDocId: string;
-  lastValidAqiTime: FirebaseFirestore.Timestamp | null;
-  lastSensorReadingTime: FirebaseFirestore.Timestamp | null;
-}
-
-/**
  * Errors that can occur during the AQI calculation process. Any of these errors
  * will result in a sensor being invalid.
  *
@@ -61,7 +32,5 @@ function getDefaultInvalidAqiErrors(): boolean[] {
   const invalidAqiErrorCount = Object.keys(InvalidAqiErrors).length / 2;
   return new Array<boolean>(invalidAqiErrorCount).fill(false);
 }
-
-export type {CurrentReadingSensorData};
 
 export {InvalidAqiErrors, getDefaultInvalidAqiErrors};
