+++ functions/src/get-reading/purple-air-response.ts
@@ -1,19 +1,11 @@
 import axios, {AxiosResponse} from 'axios';
-import {firestore, config, Timestamp, FieldValue} from '../admin';
+import {config} from '../admin';
+import {PurpleAirReading} from './types';
+import {getMeanPercentDifference} from './util';
 import {
-  populateDefaultBuffer,
-  bufferStatus,
-  Pm25BufferElement,
-  getDefaultPm25BufferElement,
-} from '../aqi-calculation/buffer';
-import {HistoricalSensorReading, PurpleAirReading} from './types';
-import {readingsSubcollection} from '../util';
-import {
-  getMeanPercentDifference,
-  getLastSensorReadingTime,
   getDefaultSensorReadingErrors,
   SensorReadingErrors,
-} from './util';
+} from './sensor-errors';
 
 /**
  * Make the PurpleAir API to using the group query
@@ -35,6 +27,7 @@ async function fetchPurpleAirResponse(): Promise<AxiosResponse> {
     'pm2.5',
     'humidity',
     'last_seen',
+    'channel_flags',
   ];
 
   // Only get readings from sensors that have an updated reading in the last 4 minutes
@@ -55,14 +48,50 @@ async function fetchPurpleAirResponse(): Promise<AxiosResponse> {
 }
 
 /**
+ * Gets any errors for channels being downgraded in a sensor
+ * @param index - index that corresponds to a flag name in the channelFlagNames
+ * @param channelFlagNames - array of flag names
+ * @returns array of errors that correspond to channel downgrade errors
+ */
+function getChannelFlag(
+  index: number,
+  channelFlagNames: string[]
+): SensorReadingErrors[] {
+  const errors: SensorReadingErrors[] = [];
+  switch (channelFlagNames[index]) {
+    case 'Normal':
+      // No channels were downgraded, so return no errors
+      break;
+    case 'A-Downgraded':
+      errors.push(SensorReadingErrors.ChannelADowngraded);
+      break;
+    case 'B-Downgraded':
+      errors.push(SensorReadingErrors.ChannelBDowngraded);
+      break;
+    case 'A+B-Downgraded':
+      errors.push(
+        SensorReadingErrors.ChannelADowngraded,
+        SensorReadingErrors.ChannelBDowngraded
+      );
+      break;
+    default:
+      // Unknown value, ignore
+      break;
+  }
+  return errors;
+}
+
+/**
  * Converts a PurpleAir reading returned from the group API query into a PurpleAirReading
  * @param data - list of data from PurpleAir for a given sensor
  * @param fieldNames - list of field names from PurpleAir that match the order of the data fields
+ * @param channelFlagNames - list of channel flag names from PurpleAir that correspond to an index in the sensor reading's data
  * @returns
  */
 function getReading(
   data: (string | number)[],
-  fieldNames: string[]
+  fieldNames: string[],
+  channelFlagNames: string[]
 ): [number, [PurpleAirReading | null, boolean[]]] {
   // Initialize all values
   let id: number = Number.NaN;
@@ -111,6 +140,14 @@ function getReading(
           timestamp = new Date(value * 1000); // eslint-disable-line no-magic-numbers
         }
         break;
+      case 'channel_flags':
+        if (typeof value === 'number') {
+          const channelDownErrors = getChannelFlag(value, channelFlagNames);
+
+          // Add the channel down errors to the overall sensor errors
+          channelDownErrors.forEach(error => (sensorErrors[error] = true));
+        }
+        break;
       default:
         // Unknown field, ignore
         break;
@@ -139,25 +176,46 @@ function getReading(
       timestamp: timestamp,
     };
 
+    /**
+     * Sometimes PurpleAir will give a non-zero confidence value when a sensor
+     * has a channel downgraded. This is not the behavior we want because we
+     * need both channels' readings to be able to determine if we want to use
+     * that reading in our AQI calculation since we want to be able to know how
+     * confident we can be that a sensor reading is true and not an outlier.
+     * Thus, if either channel of a sensor is downgraded, we set the
+     * meanPercentDifference to the maximum value of 2.
+     */
+    if (
+      sensorErrors[SensorReadingErrors.ChannelADowngraded] ||
+      sensorErrors[SensorReadingErrors.ChannelADowngraded]
+    ) {
+      const maxPercentDifference = 2;
+      meanPercentDifference = maxPercentDifference;
+    }
+
+    // If the mean percent difference is higher than the EPA threshold of 0.7,
+    // then we signal that the most recent sensor reading's channels had diverged.
     const percentDifferenceThreshold = 0.7;
     if (meanPercentDifference > percentDifferenceThreshold) {
       sensorErrors[SensorReadingErrors.ChannelsDiverged] = true;
     }
+
     return [id, [reading, sensorErrors]];
   } else {
     if (humidity === undefined) {
       sensorErrors[SensorReadingErrors.NoHumidityReading] = true;
-      sensorErrors[SensorReadingErrors.IncompleteSensorReading] = true;
-    } else {
-      // TODO: Add channel down check
     }
+
+    // If some field was undefined, then the sensor reading was incomplete
+    sensorErrors[SensorReadingErrors.IncompleteSensorReading] = true;
+
     return [id, [null, sensorErrors]];
   }
 }
 
 /**
- * Translates PurpleAir response into map of sensor ID to PurpleAirReading
- * @returns map of sensor ID to PurpleAirReading
+ * Translates PurpleAir response into map of sensor ID to PurpleAirReading and an array of sensor errors.
+ * @returns map of sensor ID to a tuple of a PurpleAirReading and a list of sensor errors. If the reading received from PurpleAir is incomplete, it is then `null`.
  */
 async function getReadingsMap(): Promise<
   Map<number, [PurpleAirReading | null, boolean[]]>
@@ -166,136 +224,13 @@ async function getReadingsMap(): Promise<
   const readings: Map<number, [PurpleAirReading | null, boolean[]]> = new Map();
   const purpleAirData = purpleAirResponse.data;
   const fieldNames: string[] = purpleAirData.fields;
+  const channelFlagNames: string[] = purpleAirData.channel_flags;
   const rawReadings: (string | number)[][] = purpleAirData.data;
   rawReadings.forEach(rawReading => {
-    const [id, reading] = getReading(rawReading, fieldNames);
+    const [id, reading] = getReading(rawReading, fieldNames, channelFlagNames);
     readings.set(id, reading);
   });
   return readings;
 }
 
-/**
- * Handles the PurpleAir API call for all active sensors,
- * storing the new data in each sensors' readings collection.
- */
-async function purpleAirToFirestore(): Promise<void> {
-  const readingsMap = await getReadingsMap();
-
-  // Add each of the new readings to the readings subcollection and the pm25buffers
-  const activeSensorDocsSnapshot = await firestore
-    .collection('sensors')
-    .where('isActive', '==', true)
-    .get();
-
-  for (const sensorDoc of activeSensorDocsSnapshot.docs) {
-    const readingsCollectionRef = firestore.collection(
-      readingsSubcollection(sensorDoc.id)
-    );
-
-    // Get the existing data from Firestore for this sensor
-    const sensorDocData: FirebaseFirestore.DocumentData =
-      sensorDoc.data() ?? {};
-    const purpleAirId: number = sensorDocData.purpleAirId;
-
-    // If the lastSensorReadingTime field isn't set, query the readings
-    // collection to find the timestamp of the most recent reading.
-    const lastSensorReadingTime: FirebaseFirestore.Timestamp | null =
-      sensorDocData.lastSensorReadingTime ??
-      (await getLastSensorReadingTime(readingsCollectionRef));
-
-    // Initialize sensor errors
-    let errors: boolean[] = getDefaultSensorReadingErrors();
-    let reading: PurpleAirReading | null = null;
-
-    // If a reading for this sensor was not in the group query, then it did not
-    // receive a new reading recently enough
-    const purpleAirResult = readingsMap.get(purpleAirId);
-
-    if (typeof purpleAirResult === 'undefined') {
-      // No reading was received from PurpleAir
-      errors[SensorReadingErrors.ReadingNotReceived] = true;
-    } else {
-      // If the reading was invalid or incomplete, the errors will be propagated
-      [reading, errors] = purpleAirResult;
-    }
-
-    const readingTimestamp: FirebaseFirestore.Timestamp | null = reading
-      ? Timestamp.fromDate(reading.timestamp)
-      : null;
-
-    // Initialize the buffer element to the default value
-    let pm25BufferElement: Pm25BufferElement = getDefaultPm25BufferElement();
-
-    // Initialize the sensor doc update data
-    const sensorDocUpdate = Object.create(null);
-    sensorDocUpdate.lastUpdated = FieldValue.serverTimestamp();
-    sensorDocUpdate.sensorReadingErrors = errors;
-
-    if (reading && readingTimestamp) {
-      // Before adding the reading to the historical database, check that it
-      // doesn't already exist in the database
-      if (
-        lastSensorReadingTime === null ||
-        lastSensorReadingTime.seconds !== readingTimestamp.seconds
-      ) {
-        // Update the buffer element from the default element
-        pm25BufferElement = {
-          timestamp: readingTimestamp,
-          pm25: reading.pm25,
-          meanPercentDifference: reading.meanPercentDifference,
-          humidity: reading.humidity,
-        };
-
-        // Update the sensor doc's data
-        sensorDocUpdate.lastSensorReadingTime = readingTimestamp;
-        sensorDocUpdate.latitude = reading.latitude;
-        sensorDocUpdate.longitude = reading.longitude;
-        sensorDocUpdate.name = reading.name;
-
-        // Add to historical readings
-        const historicalSensorReading: HistoricalSensorReading = {
-          timestamp: readingTimestamp,
-          pm25: reading.pm25,
-          meanPercentDifference: reading.meanPercentDifference,
-          humidity: reading.humidity,
-          latitude: reading.latitude,
-          longitude: reading.longitude,
-        };
-        await readingsCollectionRef.add(historicalSensorReading);
-      }
-    }
-
-    // Update the PM2.5 buffer
-    const status = sensorDocData.pm25BufferStatus ?? bufferStatus.DoesNotExist;
-    if (status === bufferStatus.Exists) {
-      // If the buffer exists, update normally
-      const pm25Buffer = sensorDocData.pm25Buffer;
-      pm25Buffer[sensorDocData.pm25BufferIndex] = pm25BufferElement;
-
-      sensorDocUpdate.pm25BufferIndex =
-        (sensorDocData.pm25BufferIndex + 1) % pm25Buffer.length; // eslint-disable-line no-magic-numbers
-      sensorDocUpdate.pm25Buffer = pm25Buffer;
-    } else if (status === bufferStatus.DoesNotExist) {
-      // Initialize populating the buffer with default values, don't update
-      // any values until the buffer status is Exists
-      sensorDocUpdate.pm25BufferStatus = bufferStatus.InProgress;
-    }
-
-    // Send the updated data to the database
-    await firestore
-      .collection('sensors')
-      .doc(sensorDoc.id)
-      .update(sensorDocUpdate);
-
-    // If the buffer didn't exist, use another write to initialize the buffer.
-    // Since the buffer is large, this can be timely and this function ensures
-    // that the buffer is not re-created while the buffer is being created.
-    if (status === bufferStatus.DoesNotExist) {
-      // This function updates the bufferStatus once the buffer has been
-      // fully initialized, which uses an additional write to the database
-      populateDefaultBuffer(false, sensorDoc.id);
-    }
-  }
-}
-
-export {purpleAirToFirestore};
+export {getReadingsMap};
