+++ functions/src/get-reading/util.ts
@@ -27,16 +27,18 @@
 function getMeanPercentDifference(confidence: number): number {
   const minConfidence = 0;
   const maxConfidence = 100;
-  /* eslint-disable no-magic-numbers */
+
+  const minPercentDifference = 0;
+  const maxPercentDifference = 2;
   switch (confidence) {
     case minConfidence:
       // If the confidence is zero, then we return the maximum possible percent difference
-      return 2;
+      return maxPercentDifference;
     case maxConfidence:
       // The confidence value from PurpleAir can be 100 even if channel A and
       // channel B do not completely match, but if the confidence value is 100,
       // then the value is data is good enough to meet the EPA recommendation.
-      return 0;
+      return minPercentDifference;
     default:
       // Otherwise, undo the calculation from the PurpleAir confidence value
       // eslint-disable-next-line no-magic-numbers
@@ -67,50 +69,4 @@ async function getLastSensorReadingTime(
   return lastSensorReadingTime;
 }
 
-/**
- * Errors that can occur with a single sensor reading. Since these errors only
- * affect a given sensor reading, these errors can occur and a sensor can still
- * have a valid AQI if enough of the other readings are valid.
- *
- * - `ReadingNotReceived` - In the most recent recent call to PurpleAir, no
- *   reading was received for this sensor.
- * - `NoHumidityReading` - No humidity reading was received for a sensor that
- *   had a reading from PurpleAir. For this error, `IncompleteSensorReading` is
- *   also true.
- * - `IncompleteSensorReading` - Some value for a sensor reading was not received.
- * - `ChannelsDiverged` - Channel A and Channel B of a PurpleAir sensor have
- *   diverged enough that the most recent sensor reading should not be used.
- *   We determine if the channels have diverged by calculating the mean percent
- *   difference between the readings for the channels from the confidence value
- *   from PurpleAir. The EPA recommends using a maximum mean percent difference
- *   of 0.7 for valid sensor readings.
- * - `ChannelADowngraded` - If PurpleAir has downgraded Channel A of a sensor.
- * - `ChannelBDowngraded` - If PurpleAir has downgraded Channel B of a sensor.
- */
-enum SensorReadingErrors {
-  ReadingNotReceived,
-  NoHumidityReading,
-  IncompleteSensorReading,
-  ChannelsDiverged,
-  ChannelADowngraded,
-  ChannelBDowngraded,
-}
-
-/**
- * Returns a default error array for sensor errors
- * @returns an array for each `SensorReadingError` with the error set to `false`.
- */
-function getDefaultSensorReadingErrors(): boolean[] {
-  // TypeScript does not provide a way to get the number of elements in an
-  // enumeration, so this gets the number of elements using the enumeration
-  // reverse mapping.
-  const sensorReadingErrorCount = Object.keys(SensorReadingErrors).length / 2;
-  return new Array<boolean>(sensorReadingErrorCount).fill(false);
-}
-
-export {
-  getLastSensorReadingTime,
-  getMeanPercentDifference,
-  SensorReadingErrors,
-  getDefaultSensorReadingErrors,
-};
+export {getLastSensorReadingTime, getMeanPercentDifference};
